Tried to improve the constructors.
The two specific constructor (for markdown and mediawiki) were using the same properties.

I moved it to the parent class, and just called the superclass constructor in the two previous constructors, passing their specific values.

However this does not work, dur to another issue with the OOP library: the "optional new" feature seems to break it. Indeed, the parent doesn't work on the actual this, but creates its new one, maybe because it thinks it's not from its inheritance chain.